backend/requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
cssselect2==0.7.0
Django==4.1.3
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-ipware==4.0.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
html5lib==1.1
idna==3.4
kombu==5.2.4
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
PyPDF3==1.0.6
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.4.2
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
ua-parser==0.16.1
uritools==4.0.1
urllib3==1.26.14
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.8
